auto logged in for dev server
